unifying JacobianOperator in ShellUtilities
