+++ applications/StructuralMechanicsApplication/custom_utilities/shell_utilities.h
@@ -24,10 +24,56 @@
 #include "containers/array_1d.h"
 
 namespace Kratos {
+
+class ShellQ4_LocalCoordinateSystem; // forward declare to avoid circular includes
+
 namespace ShellUtilities {
 
 using SizeType = std::size_t;
 using IndexType = std::size_t;
+
+/** \brief JacobianOperator
+ *
+ * This class is a utility to compute at a given integration point,
+ * the Jacobian, its inverse, its determinant
+ * and the derivatives of the shape functions in the local
+ * cartesian coordinate system.
+ */
+class JacobianOperator
+{
+public:
+
+    JacobianOperator();
+
+    void Calculate(const ShellQ4_LocalCoordinateSystem& CS, const Matrix& dN);
+
+    inline const Matrix& Jacobian()const
+    {
+        return mJac;
+    }
+
+    inline const Matrix& Inverse()const
+    {
+        return mInv;
+    }
+
+    inline const Matrix& XYDerivatives()const
+    {
+        return mXYDeriv;
+    }
+
+    inline double Determinant()const
+    {
+        return mDet;
+    }
+
+private:
+
+    Matrix mJac;     /*!< Jacobian matrix */
+    Matrix mInv;     /*!< Inverse of the Jacobian matrix */
+    Matrix mXYDeriv; /*!< Shape function derivatives in cartesian coordinates */
+    double mDet;     /*!< Determinant of the Jacobian matrix */
+};
 
 template<class TVec>
 inline void ShapeFunc(double Xi, double Eta, TVec& rN)
